Use layout file and move AppShell into it
