+++ app/(parts)/layout.tsx
@@ -1,68 +1,16 @@
 "use client";
 
-import PartsTab from "@/components/PartsTab";
 import { ThemeDropdown } from "@/components/ThemeDropdown";
 import TotalPrice from "@/components/TotalPrice";
-import { Cpu as CpuEntity } from "@/db/Cpu";
 import pages from "@/utils/pages.json";
 import { Anchor, AppShell, Burger, Group, NavLink } from "@mantine/core";
 import { useDisclosure } from "@mantine/hooks";
 import Link from "next/link";
-import { useEffect, useState } from "react";
-import "reflect-metadata";
-import initSqlJs from "sql.js";
-import { Fetcher } from "swr";
-import useSWRImmutable from "swr/immutable";
-import { DataSource } from "typeorm";
+import { usePathname } from "next/navigation";
 
-const fetcher: Fetcher<ArrayBuffer, string> = (url) =>
-  fetch(url).then((res) => res.arrayBuffer());
-
-function useBuf() {
-  const { data } = useSWRImmutable(
-    "https://bucket.nunawa.net/parts_latest.db",
-    fetcher,
-  );
-
-  return {
-    buf: data,
-  };
-}
-
-const pageList = pages;
-
-export default function Cpu() {
-  const { buf } = useBuf();
-  const [sql, setSql] = useState<initSqlJs.SqlJsStatic>();
-  const [dataSource, setDatasource] = useState<DataSource>();
+export default function Page({ children }: { children: React.ReactNode }) {
   const [opened, { toggle }] = useDisclosure();
-
-  useEffect(() => {
-    initSqlJs({
-      locateFile: () =>
-        new URL("sql.js/dist/sql-wasm.wasm", import.meta.url).toString(),
-    }).then((sql) => {
-      setSql(sql);
-    });
-  }, []);
-
-  useEffect(() => {
-    if (buf && sql) {
-      (async () => {
-        globalThis.SQL = sql;
-
-        const dataSource = new DataSource({
-          type: "sqljs",
-          entities: [CpuEntity],
-          database: new Uint8Array(buf),
-        });
-
-        await dataSource.initialize();
-
-        setDatasource(dataSource);
-      })();
-    }
-  }, [buf, sql]);
+  const pathname = usePathname();
 
   return (
     <AppShell
@@ -98,8 +46,8 @@ export default function Cpu() {
         </Group>
       </AppShell.Header>
       <AppShell.Navbar p="md">
-        {pageList.map((x) => {
-          if (x.key === "cpu") {
+        {pages.map((x) => {
+          if (x.key === pathname.slice(1)) {
             return <NavLink key={x.key} label={x.name} active />;
           } else {
             return (
@@ -119,9 +67,7 @@ export default function Cpu() {
           href="/build"
         />
       </AppShell.Navbar>
-      <AppShell.Main>
-        <PartsTab type="cpu" dataSource={dataSource} />
-      </AppShell.Main>
+      <AppShell.Main>{children}</AppShell.Main>
     </AppShell>
   );
 }
